mainTopBanner api 구현

+++ app/src/main/java/com/example/planet/data/dto/Advertisement.kt
@@ -1,7 +1,7 @@
 package com.example.planet.data.dto
 
 data class Advertisement(
-    val advertisementId: Int,
+    val advertisementId: Long,
     val imageUrl: String,
     val linkUrl: String
 )
